Add server side page state handling
I created a function getPageState that should be on page component.
On the server when the page is request, the page component calls the
getPageState function and resolves promises that should request data.
The page is rendered with initial data structured like object returned
from getPageState.

getInitialState resolves and maps the resolved data to the provided
strcture.

getInitialHTML exposes initial state on global level and renders the
component on the server. Initial state is available on the server and
the client.